requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
docutils==0.15.2
idna==2.8
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
Pygments==2.4.2
pyparsing==2.4.2
pyserial==3.4
python-dateutil==2.8.0
requests==2.22.0
scipy==1.3.1
six==1.12.0
sortedcontainers==2.1.0
urllib3==1.25.3
